Fixes #59 (#60)

+++ src/marlin/gcode/gcode.h
@@ -998,7 +998,7 @@ private:
 
 	TERN_(CONTROLLER_FAN_EDITABLE, static void M710());
 
-	static void T(const int8_t tool_index);
+	static void T(const int16_t tool_index);
 
 public:
 	static const char* get_state();
